# client/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Weather API",
    "version": "v1"
  },
  "paths": {
...
    "/api/WeatherReading": {
      "get": {
        "tags": [
          "WeatherReading"
        ],
        "operationId": "GetReadings",
        "parameters": [
          {
            "name": "Start",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "End",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "StationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "PageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoListApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoListApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoListApiResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "WeatherReading"
        ],
        "operationId": "CreateReading",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeatherReadingRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeatherReadingRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWeatherReadingRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/WeatherReading/{id}": {
      "get": {
        "tags": [
          "WeatherReading"
        ],
        "operationId": "GetReadingById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/WeatherReadingDtoApiResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "WeatherReading"
        ],
        "operationId": "DeleteReading",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateWeatherReadingRequest": {
        "type": "object",
        "properties": {
          "stationId": {
            "type": "string",
            "format": "uuid"
          },
          "mainCondition": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "double"
          },
          "feelsLike": {
            "type": "number",
            "format": "double"
          },
          "minTemp": {
            "type": "number",
            "format": "double"
          },
          "maxTemp": {
            "type": "number",
            "format": "double"
          },
          "pressure": {
            "type": "integer",
            "format": "int32"
          },
          "humidity": {
            "type": "integer",
            "format": "int32"
          },
          "seaLevel": {
            "type": "integer",
            "format": "int32"
          },
          "groundLevel": {
            "type": "integer",
            "format": "int32"
          },
          "visibility": {
            "type": "integer",
            "format": "int32"
          },
          "windSpeed": {
            "type": "number",
            "format": "double"
          },
          "windDeg": {
            "type": "integer",
            "format": "int32"
          },
          "cloudiness": {
            "type": "integer",
            "format": "int32"
          },
          "rain": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "snow": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
      "WeatherReadingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "stationId": {
            "type": "string",
            "format": "uuid"
          },
          "mainCondition": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "icon": {
            "type": "string",
            "nullable": true
          },
          "temperature": {
            "type": "number",
            "format": "double"
          },
          "feelsLike": {
            "type": "number",
            "format": "double"
          },
          "minTemp": {
            "type": "number",
            "format": "double"
          },
          "maxTemp": {
            "type": "number",
            "format": "double"
          },
          "pressure": {
            "type": "integer",
            "format": "int32"
          },
          "humidity": {
            "type": "integer",
            "format": "int32"
          },
          "seaLevel": {
            "type": "integer",
            "format": "int32"
          },
          "groundLevel": {
            "type": "integer",
            "format": "int32"
          },
          "visibility": {
            "type": "integer",
            "format": "int32"
          },
          "windSpeed": {
            "type": "number",
            "format": "double"
          },
          "windDeg": {
            "type": "number",
            "format": "double"
          },
          "cloudiness": {
            "type": "integer",
            "format": "int32"
          },
          "rain": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "snow": {
            "type": "number",
            "format": "double",
            "nullable": true
          },
          "capturedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "WeatherReadingDtoApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "$ref": "#/components/schemas/WeatherReadingDto"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "WeatherReadingDtoListApiResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WeatherReadingDto"
            },
            "nullable": true
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "Enter JWT Bearer token **_only_**",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}